trolldirectory/requirements.txt
absl-py~=1.0.0
astunparse~=1.6.3
cachetools~=5.1.0
certifi~=2022.5.18.1
charset-normalizer~=2.0.12
cvlib~=0.2.7
flatbuffers~=1.12
gast~=0.4.0
google-auth~=2.6.6
google-auth-oauthlib~=0.4.6
google-pasta~=0.2.0
grpcio~=1.46.1
h5py~=3.6.0
idna~=3.3
imageio~=2.19.2
imutils~=0.5.4
keras~=2.9.0
Keras-Preprocessing~=1.1.2
libclang~=14.0.1
Markdown~=3.3.7
numpy~=1.22.3
oauthlib~=3.2.0
opencv-python~=4.5.5.64
opt-einsum~=3.3.0
packaging~=21.3
Pillow~=9.1.1
progressbar~=2.5
protobuf~=3.20.1
pyasn1~=0.4.8
pyasn1-modules~=0.2.8
pyparsing~=3.0.9
requests~=2.27.1
requests-oauthlib~=1.3.1
rsa~=4.8
six~=1.16.0
tensorboard~=2.9.0
tensorboard-data-server~=0.6.1
tensorboard-plugin-wit~=1.8.1
tensorflow~=2.9.0
tensorflow-estimator~=2.9.0
tensorflow-io-gcs-filesystem~=0.26.0
termcolor~=1.1.0
typing_extensions~=4.2.0
urllib3~=1.26.9
Werkzeug~=2.1.2
wrapt~=1.14.1